+++ NEWS.md
@@ -7,6 +7,8 @@
 * fixed bug where most data was ignored (this was because of non-matching character types in fields of the input data)
 * fixed issue with pandoc not being found (NOTE: this may not work on other computers)
 * fixed conflict betwwen normal and excel modes about where save files were placed 
+* document now correctly states whether inflation was applied
+* added installation of DfeR to install instructions as it is necessary for the document template.